moved sQP to queue-proxy/
