+++ tests/go.mod
@@ -8,7 +8,7 @@ replace (
 	github.com/ease-lab/xdt/proto/crossXDT => ../proto/crossXDT
 	github.com/ease-lab/xdt/proto/downXDT => ../proto/downXDT
 	github.com/ease-lab/xdt/proto/upXDT => ../proto/upXDT
-	github.com/ease-lab/xdt/sQP => ../sQP
+	github.com/ease-lab/xdt/queue-proxy/sQP => ../queue-proxy/sQP
 	github.com/ease-lab/xdt/sdk/go_sdk => ../sdk/go_sdk
 	github.com/ease-lab/xdt/transport => ../transport
 	github.com/ease-lab/xdt/utils => ../utils
@@ -19,7 +19,7 @@ require (
 	github.com/containerd/containerd v1.5.2
 	github.com/ease-lab/vhive/utils/tracing/go v0.0.0-20210624210547-e0cd5d053491
 	github.com/ease-lab/xdt/dQP v0.0.0-00010101000000-000000000000
-	github.com/ease-lab/xdt/sQP v0.0.0-00010101000000-000000000000
+	github.com/ease-lab/xdt/queue-proxy/sQP v0.0.0-00010101000000-000000000000
 	github.com/ease-lab/xdt/sdk/go_sdk v0.0.0-00010101000000-000000000000
 	github.com/ease-lab/xdt/utils v0.0.0-00010101000000-000000000000
 	github.com/sirupsen/logrus v1.8.1
