+++ tests/unit_test.go
@@ -34,7 +34,7 @@ import (
 	ctrdlog "github.com/containerd/containerd/log"
 
 	"github.com/ease-lab/xdt/dQP"
-	"github.com/ease-lab/xdt/sQP"
+	"github.com/ease-lab/xdt/queue-proxy/sQP"
 	"github.com/ease-lab/xdt/utils"
 
 	log "github.com/sirupsen/logrus"
