+++ user-functions/sQP/go.mod
@@ -5,7 +5,7 @@ go 1.16
 replace (
 	github.com/ease-lab/xdt/proto/crossXDT => ../../proto/crossXDT
 	github.com/ease-lab/xdt/proto/upXDT => ../../proto/upXDT
-	github.com/ease-lab/xdt/sQP => ../../sQP
+	github.com/ease-lab/xdt/queue-proxy/sQP => ../../queue-proxy/sQP
 	github.com/ease-lab/xdt/transport => ../../transport
 	github.com/ease-lab/xdt/utils => ../../utils
 )
@@ -13,7 +13,7 @@ replace (
 require (
 	github.com/containerd/containerd v1.5.2
 	github.com/ease-lab/vhive/utils/tracing/go v0.0.0-20210623161727-460bac97d8c0
-	github.com/ease-lab/xdt/sQP v0.0.0-00010101000000-000000000000
+	github.com/ease-lab/xdt/queue-proxy/sQP v0.0.0-00010101000000-000000000000
 	github.com/ease-lab/xdt/utils v0.0.0-00010101000000-000000000000
 	github.com/sirupsen/logrus v1.8.1
 )
